attempt to send data to HMI
set up HMI screen (Just a text box). In microchip studio, try to set up USART. there should be a counter displayed on the HMI, but nothing is being sent. some problems may be the incorrect (overflow or wrong val) BAUDRATE, using the wrong ports, connecting to the wrong ports